Rename AdminPage to AdminStockPage
