+++ src/pages/admin/AdminStockPage/element.ts
