+++ src/pages/admin/AdminStockPage/index.ts
